requirements_base.txt
otree>=2.6.0b16
djangorestframework